updated location of toPascalCase unit tests
