+++ apps/vs-code-designer/src/app/utils/codeful/unitTestUtils.test.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
 import { extractAndValidateRunId, removeInvalidCharacters, validateRunId } from '../unitTests';
-import { toPascalCase } from '@microsoft/logic-apps-shared';
 
 // Mock the vscode-nls module
 vi.mock('vscode-nls', () => ({
@@ -76,60 +75,5 @@ describe('removeInvalidCharacters', () => {
     const input = '';
     const result = removeInvalidCharacters(input);
     expect(result).toBe('');
-  });
-});
-
-describe('toPascalCase', () => {
-  it('should convert a valid snake_case string to PascalCase', () => {
-    const input = 'example_string';
-    const result = toPascalCase(input);
-    expect(result).toBe('ExampleString');
-  });
-
-  it('should capitalize the first letter of a single word', () => {
-    const input = 'example';
-    const result = toPascalCase(input);
-    expect(result).toBe('Example');
-  });
-
-  it('should handle strings with multiple underscores', () => {
-    const input = 'example__string__test';
-    const result = toPascalCase(input);
-    expect(result).toBe('ExampleStringTest');
-  });
-
-  it('should handle an empty string gracefully', () => {
-    const input = '';
-    const result = toPascalCase(input);
-    expect(result).toBe('');
-  });
-
-  it('should handle strings with leading underscores', () => {
-    const input = '_example_string';
-    const result = toPascalCase(input);
-    expect(result).toBe('ExampleString');
-  });
-});
-
-describe('combined usage of removeInvalidCharacters and toPascalCase', () => {
-  it('should clean and convert a string to PascalCase', () => {
-    const input = 'example-string(123)';
-    const cleaned = removeInvalidCharacters(input);
-    const pascalCase = toPascalCase(cleaned);
-    expect(pascalCase).toBe('ExampleString123');
-  });
-
-  it('should handle strings with only invalid characters', () => {
-    const input = '!@#$%^&*()';
-    const cleaned = removeInvalidCharacters(input);
-    const pascalCase = toPascalCase(cleaned);
-    expect(pascalCase).toBe('');
-  });
-
-  it('should handle already clean PascalCase strings', () => {
-    const input = 'ExampleString';
-    const cleaned = removeInvalidCharacters(input);
-    const pascalCase = toPascalCase(cleaned);
-    expect(pascalCase).toBe('ExampleString');
   });
 });
