+++ libs/logic-apps-shared/src/utils/src/lib/helpers/__test__/stringFunctions.spec.ts
@@ -1,4 +1,4 @@
-import { escapeString, idDisplayCase, labelCase, canStringBeConverted, unescapeString } from '../stringFunctions';
+import { escapeString, idDisplayCase, labelCase, canStringBeConverted, unescapeString, toPascalCase } from '../stringFunctions';
 import { describe, it, expect } from 'vitest';
 describe('label_case', () => {
   it('should replace _ with spaces', () => {
@@ -155,5 +155,37 @@ describe('escapeString', () => {
 
   it('should handle an empty string', () => {
     expect(escapeString('')).toEqual('');
+  });
+});
+
+describe('toPascalCase', () => {
+  it('should convert a valid snake_case string to PascalCase', () => {
+    const input = 'example_string';
+    const result = toPascalCase(input);
+    expect(result).toBe('ExampleString');
+  });
+
+  it('should capitalize the first letter of a single word', () => {
+    const input = 'example';
+    const result = toPascalCase(input);
+    expect(result).toBe('Example');
+  });
+
+  it('should handle strings with multiple underscores', () => {
+    const input = 'example__string__test';
+    const result = toPascalCase(input);
+    expect(result).toBe('ExampleStringTest');
+  });
+
+  it('should handle an empty string gracefully', () => {
+    const input = '';
+    const result = toPascalCase(input);
+    expect(result).toBe('');
+  });
+
+  it('should handle strings with leading underscores', () => {
+    const input = '_example_string';
+    const result = toPascalCase(input);
+    expect(result).toBe('ExampleString');
   });
 });
